notifications_admin/requirements.txt
Django==4.1.13
django-debug-toolbar==4.3.0
gunicorn==22.0.0
psycopg2==2.9.10
celery==5.5.2
redis==5.0.4
django-celery-beat==2.4.0
django-redis==5.2.0
django-allauth==0.54.0
split-settings==1.0.0
sentry-sdk[django]==2.27.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0